updated: Resource naming to Result and moved to common module instead of network
