+++ feature/details/build.gradle.kts
@@ -26,6 +26,5 @@ android.namespace = "dev.enesky.feature.details"
 dependencies {
     implementation(projects.core.common)
     implementation(projects.core.domain)
-    implementation(projects.core.network)
     implementation(projects.core.designSystem)
 }
